Refactor news-card according to designs
