+++ front-end/src/components/newsfeed-view/news-list/card-group/news-card-group.module.scss
@@ -1,0 +1,4 @@
+.news-group {
+    display: flex;
+    flex-direction: column;
+}
