+++ front-end/src/components/newsfeed-view/news-list/card-group/news-card/InstagramCard.js
@@ -1,19 +1,20 @@
 import React from 'react';
 import Timestamp from './elements/Timestamp';
-import styles from './news-card.module.scss';
+import styles from './social-media-card.module.scss';
 
-const TwitterNewsCard = ({data}) => (
-    <div className={`${styles['news-card']} ${styles['vertical']} ${styles['twitter']}`}>
+const InstagramCard = ({data}) => (
+    <div className={`${styles['news-card']} ${styles['vertical']} ${styles['instagram']}`}>
         <div className={styles['content']}>
             <div className={styles['source']}>
                 <div className={styles['author']}>
-                    <img src = {data.thumbnail} />
-                    <p>{data.author}</p>
+                <img src = {data.thumbnail} />
+                 <p>{data.author}</p>
                 </div>
                 <span>{data.source}</span>
             </div>
+            <div className={styles['image']} style={{ backgroundImage: `url(${data.img})` }}></div>
             <div className={styles['description']}>
-                <p>{data.description}</p>
+                <p><b>{data.author}</b> {data.description}</p>
             </div>
             <div className={styles['footer']}>
                 <Timestamp time = {data.timestamp} />
@@ -22,4 +23,4 @@ const TwitterNewsCard = ({data}) => (
     </div>
  );
 
- export default TwitterNewsCard;
+ export default InstagramCard;
