+++ front-end/src/components/newsfeed-view/news-list/card-group/news-card/NewsCard.js
@@ -1,18 +1,24 @@
 import React from 'react';
 import Timestamp from './elements/Timestamp';
 import styles from './news-card.module.scss';
-
+import LineEllipsis from 'react-lines-ellipsis';
 
 const NewsCard = ({data}) => (
     <div className={styles["news-card"]}>
+        <div className={styles["source"]}>
+            {data.source}
+        </div>
         <div className={styles["image"]} style={{ backgroundImage: `url(${data.img})` }}></div>
         <div className={styles["content"]}>
-            <div className={styles["source"]}>
-                <span>{data.source}</span>
-            </div>
             <div className={styles["description"]}>
-                <p>{data.title}</p>
-                <span>by @authorname</span>
+                <span>Author / page headline</span>
+                <p>
+                    <LineEllipsis
+                        text={data.title}
+                        maxLine='3'
+                    >
+                    </LineEllipsis>
+                </p>
             </div>
             <div className={styles["footer"]}>
                 <Timestamp time={data.timestamp} />
