+++ front-end/src/components/newsfeed-view/news-list/card-group/news-card/TwitterCard.js
@@ -1,20 +1,19 @@
 import React from 'react';
 import Timestamp from './elements/Timestamp';
-import styles from './news-card.module.scss';
+import styles from './social-media-card.module.scss';
 
-const InstagramNewsCard = ({data}) => (
-    <div className={`${styles['news-card']} ${styles['vertical']} ${styles['instagram']}`}>
+const TwitterCard = ({data}) => (
+    <div className={`${styles['news-card']} ${styles['vertical']} ${styles['twitter']}`}>
         <div className={styles['content']}>
             <div className={styles['source']}>
                 <div className={styles['author']}>
-                <img src = {data.thumbnail} />
-                 <p>{data.author}</p>
+                    <img src = {data.thumbnail} />
+                    <p>{data.author}</p>
                 </div>
                 <span>{data.source}</span>
             </div>
-            <div className={styles['image']} style={{ backgroundImage: `url(${data.img})` }}></div>
             <div className={styles['description']}>
-                <p><b>{data.author}</b> {data.description}</p>
+                <p>{data.description}</p>
             </div>
             <div className={styles['footer']}>
                 <Timestamp time = {data.timestamp} />
@@ -23,4 +22,4 @@ const InstagramNewsCard = ({data}) => (
     </div>
  );
 
- export default InstagramNewsCard;
+ export default TwitterCard;
